requirements.txt
Flask_Cors==3.0.8
torchvision==0.4.0
Flask==1.1.1
https://download.pytorch.org/whl/cpu/torch-1.2.0%2Bcpu-cp36-cp36m-manylinux1_x86_64.whl
numpy==1.16.0
boto3==1.9.215
Pillow==6.1.0
gunicorn==19.9.0
requests==2.21.0